Fix #5486
Sort now treats None and UNDEFINED_DATE as UNDEFINED_DATE

It is now possible to set a date to None while editing directly on the library screen. To do so, set the date to 1/1/101.

The metadata editors show 'Undefined' for None dates. However, when a date widget gets focus, the display changes to showing the 'real' UNDEFINED_DATE. When the focus leaves, the widget reverts, unless it has been changed.