No device collections if shelves not maintained
Kobo driver: Display device collections even if the driver is configured
to not manage shelves on the device. Fixes #1214233 [Collections are empty when not maintaining shelves on Kobo devices](https://bugs.launchpad.net/calibre/+bug/1214233)

Skip getting the shelf collections if the Kobo driver is not configured
to maintain them. This was wiping out the device collections.