Update Kobo firmware version. AGAIN.
Kobo are updating the firmware a lot faster than in the past. The latest
is 3.1.0. The supported version is being set to 3.1.5 in case any bug
fix only versions are released.

Also, commented out a couple of debug lines to reduce the clutter when
testing.

Fixes #1257143 [Update to support Kobo firmware 3.1.0](https://bugs.launchpad.net/calibre/+bug/1257143)