Fix content server internal redirects not unquoting the category.
See http://www.mobileread.com/forums/showthread.php?t=236441 for some context.

Although the fix works, I am not sure that this is the right way to fix it. I am not familiar enough with the routing system to know if there is a single place somewhere that is missing the unquote.