Remove escaped EM tags, keeping only content
The source epub contains multiple escaped EM tags, also in title tags
and the table of contents. The safest way (and easiest way) to get a
readable epub from this is to dispense with the escaped EM tags
altogether, everywhere in .html and .ncx files (instead of replacing
them with unescaped ones), of course while keeping the human-readable
contents in place.