Update to support Kobo firmware 3.5.0
Kobo are about to release firmware 3.5.0. The beta and release
candidates have been tested with calibre and the version numbers is all
that needs to change. As usual, I am bumping it to 3.5.1 in case of a
second bug fix release with no backend changes.

The only other change is in the list of cover images. This has a number
for the maximum DBVersion for the image. The previous value has been
exceeded, so it is being bumped again.

Fixes #1340039 [Update to support Kobo firmware 3.5.0](https://bugs.launchpad.net/calibre/+bug/1340039)