Replace deprecated __defineGetter__ and __defineSetter___
__defineGetter__ was not working in Qt 5, this caused the following
regressions:

PDF Output: Header and footer templates not working in calibre 2.0.
Fixes #1361483 [Headers and footers don't appear EPUB to PDF](https://bugs.launchpad.net/calibre/+bug/1361483)

E-book viewer: The Table of Contents not displaying the currently viewed
section in bold if the section is a link to a location in the middle of
a file.