Kobo driver: Update to support new firmware from Kobo. Fixes #1382051 [Kobo: Skip modifying epub if no valid CSS rules and bump firmware version](https://bugs.launchpad.net/calibre/+bug/1382051)
Skip modifying epub if no valid CSS rules and bump firmware version

While looking at a problem with the stylesheet on Kobo devices recently,
I realised that if the modify CSS option was set and the was no file, it
still modified the stylesheet. Now, will not do this if there are no
valid rules in the kobo_extra.css file.

Also bumping firmware version to 3.11.1. Kobo are about to release an
update. They big jump on version number is to match the desktop
application version.