Configurable suppression of replace result dialog (see 1610658)
This is the second part of the proposal filed originally as bug #1609111.

The result dialog, shown after replacements, slows down book editing in some cases.
For example, when the user repeatedly applies his own regex-function to relatively small pre-marked text fragments.

User should have a choice to suppress the result dialog.

This implementation allows for the following switch inside the user replace function:
replace.suppress_result_dialog = True

This way it nicely fits to the first part (see #1609111) and it allows individual configuration from case to case.
Actually, it's not dependent on the other two switches ('call_after_last_match' and 'append_final_output_to_marked'),
so it can be used in combination with any of them or even individually.

This implementation also should not have any impact on other modes of work and it should not change the previous behavior.
It also should not break the backwards compatibility, I hope.