Change the application name set in Qt to calibre from libprs500. Fixes #1623397 [[Enhancement] Change APP_UID from 'libprs500' to 'Calibre'](https://bugs.launchpad.net/calibre/+bug/1623397)
I looked into the code and the only place QSettings appears to be used
is to store the last used settings of the Qt WebKit Inspector. I think
it is acceptable to lose those settings to get rid of libprs500,
especially since some X11 implementations brokenly display the app name
in their user interfaces.