Fix collections not displaying in device list for KoboTouch driver
Collections were not being shown in the device list if either the
collection management was off in the driver, or no collections column
were configured. There were two reasons, one was the empty configuration
string was handled incorrectly. The other was the device_collections
attribute could be cleared by the smart_update.