Add a tweak to allow skipping network check on news download.
The mechanism used for checking for available network connections
may not always be accurate, so allow it to be skipped. For instance
network manager may not be aware of all available connection and
so return a false negative.

This has been added as a tweak to allow those who know what they're
doing to use the alternate behavior, without breaking the default
check mechanism.