py3: make cover grid texture selection work
map() is an iterable, so using it as a shortcut for a proper for loop in
order to apply functions, is no longer valid.

Also move from the long deprecated string.*() functions -- which no
longer exist in python3 -- that are better dealt with through the native
str/unicode type methods.