macOS: Fix changing colors from dark->light not working correctly
As of Qt 5.15.0 if you set any colors in the application palette
they are not reset when the system colors are changed. So we cannot use
that to fix colors. Instead set a widget specific override palette.

Also, Qt appears to have fixed link and alternating row colors, so no
need for those override anymore.