Revert "Fix building on Arch"
This reverts commit 9b89defdb78de2a4b065178d5520deccc3567fdf.

libraqm (an obscure PIL functionality) was recently switched to be
directly linked by Pillow, rather than dlopened, which meant it was a
bit of a hidden dependency since calibre uses the imagingft module. But
it is now directly depended on by the Arch packaging, so it no longer
needs to be installed to get a functional Pillow.