fix the linux desktop filename mapping for ebook-edit
This was initially added in commit
dea0de68d2b1cf23bbd73d624588683d5b59c136 with a filler name that,
however, didn't match the usual style of desktop filenames. When the
actual desktop file was added in commit
be4b3a845b564f6f04eb908bce5fdf729beacbd4 it didn't match the filename.

Although it was mass renamed from "tweak" to "edit" in commit
97d8382a4fc13458a528ba2bf81c9a0fe9c934d3, the word ordering was still
wrong.

As a result, application menus failed to correlate open windows to
pinned icons, or (on application menus that support it) group the
windows together.