fix the linux desktop filename mapping for lrfviewer
In commit 3a889a4a634faeab1d6e4c5e988b7955e1204dde, this was added with
the wrong name -- the "lrfviewer" application doesn't have a hyphenated
name, oddly enough, even though it is in fact "ebook-viewer" not
"ebookviewer".

On the other hand, it's probably not worth fixing the inconsistency,
because changing desktop filenames can be (a bit) disruptive if people
have pinned or modified them, and it's not worth it just for a stray
hyphen. Changing the actual program executable name is even less worth
it...