Another enum cleanup.
While testing that I found several problems related to adding and editing columns when the table was sorted differently.

Also prevented moving rows when the table is sorted by anything but row number ascending. This isn't the nicest fix but making it work with new columns while the sort is descending is harder than I want to deal with tonight.