Prefs / Look & feel / Tag browser partition and hierarchy: change lists of keys to list widget with check boxes.
NB: the partition list was a gpref, which was wrong. It is now a db pref. I added migration code to the tag browser. It is possible that the server code must be changed to get the db pref. I don't know how to do that, if it is needed.