Fix a crash with a specific library
I haven't tracked down the root cause but simply updating the book
details panel after an event loop tick fixes it, so good enough for now.

Simply calling db.get_metadata() in the handler for current row changed
causes the crash. Which means it is some custom column template that is
depending on something in Qt causing the issue. There are enough
templates in this library that I dont feel like tracking it down.