Outsmart clang
clang insists INVALID_HANDLE_VALUE is not a constexpr because it is defined as (void*)-1

So it wont let us use it a template paramter. Make the template paramter
a function instead that returns the constant value. Hopefully compilers
are smart enough to just inline the function call so there is no
performance impact. Dont have the patience to check the assembly to make
sure.