Bug #2019457: Tag browser - transient right click error.
The search category can disappear from field_metadata. Perhaps other dynamic categories can as well. The implication is that dynamic categories are being removed, but how that would happen is a mystery. I suspect a plugin is operating on the "real" field_metadata instead of a copy, thereby changing the dict used by the rest of calibre.

 As it can happen, to avoid key errors check that a category exists before offering to unhide it.