build: pass on the user's compiler and compiler-flags settings to qmake
PyQt-builder uses qmake under the hood, which has a frustrating habit of
not respecting standard stuff like $CC / $CFLAGS unless the project
authors go out of their way to manually specify to use the same value as
gotten from the environment.

Implement this for the same environment variables that we currently respect
when building internal extensions *without* PyQt-builder.