Rather than recreating QTimer objects on every update use a single one per context
Also change the timeout to 100ms which is the commonly accepted
threshold for human visual response. And stop the book details panel
debounce timer during shutdown.

The debounce timers in the popups are independent and use only data
local to the popup object so probably dont need to be shutdown.