tests: fix SSL test on python 3.13 by being more lenient about VERIFY_X509_STRICT
In python 3.13, this flag was added to the default created ctx. This
seems reasonable for production validation, but the unittest code
generates a certificate that fails this. Inside test code, it seems fine
to relax the constraints again. Our goal is to test the server itself,
anyway, not the ssl module.

See also: https://github.com/python/cpython/issues/107361