tts: gracefully handle the lack of a speechd module
We already gracefully handle the lack of speechd having the
`speech-dispatcher` command utility available, and respond by disabling
that backend. Go one step further, and disable it if the module is
missing too.

It's not even the default engine. It seems fair for piper users to not
require having this installed. The build tests will assert its presence
either way so it's not like it will go unnoticed.