tests: add variable to suppress TTS tests
TTS support can be gated on e.g. Gentoo's USE=speech, so we don't want
to run the basic selfcheck of speech support if this is disabled. We do
want to run the general Qt checks, though (and also do a selfcheck of
speech support if the USE=speech is enabled!) -- so add a sneaky
environment variable that we can use to conditionally skip running those
parts.