Another try at preventing random edit dialogs while not losing "editing" when tabbing over read-only fields.
The problem of random dialogs is handled in delegates.py. Losing "editing" when tabbing is handled by jumping over fields that are readonly or are implemented by a doalog. This is a behavior change. We will see whether it will result in screams of complaint.

NB: after hours of trying, this was the only solution I found to both the random dialog and the loss of editing.