Pass the tag browser icon names to the content server client.
The name is in the dict entry 'value_icon', which might not exist, be a file name, or be None. This is a path rooted at the 'tb_icons' directory. Icons that come from templates are in the 'template_icons' subdirectory, in which case a slash is used as the path separator.