Better fix for the Qt book list phantom edits issue
The phantom edits are happening because of the mirroring with the pin
view. Apparently in some update Qt has started triggering edits on
currentChanged events. Sigh.

For the moment we disable editing of cells in the mirrored view when it
is hidden this worksaround the problem for most people, need a better
fix for when the view is being used.