Kobo driver: Possible fix for kobo renderer ignoring hyphenation
I am guessing the Kobo renderer basically ignores a <style> tag with
id=kobostylehacks as this was likely a legacy hack implemented and later
ignored by Kobo.

So put the hyphenation and other extra css into its own <style> tag.
Hopefully this will get the Kobo rendered to not ignore it.