Turn off signing of exe and dll files on windows
Now only the installers are signed. This is needed because
DigiCert now only allows a thousand files to be a signed a year.
And charges $800 per year for the privilege. There is an alternative of
using a HSM (hardware security module) but that doesnt work with
my remote VM based signing workflow.

There are over 200 exes and dlls in a single calibre install.
Until a better provider is found, dont sign them.