Refactor shelf layout code
Layout is now done in a separate thread, shelf by shelf. As each shelf
is layouted it is rendered if visible. Also fix various bugs in the old
layout code, such as not taking width of vertical scrollbar into
account.

Note that cover hover functionality is currently broken, it needs to be
refactored and changed to always shift books, to the left or the right
depending on which edge the hovered books is closer too. We might also
want to always expand the cover for the current book not just the
hovered book. Or maybe not, to be decided.