Change the default hover action back to shift
The problem with above is that it can cause neighboring books to be
completely hidden meaning you have to move the mouse right and then left
again to hover over a neighboring book. This problem is much less
pronounced in the shift case, where it tends to happen only for very
narrow books and only if one keeps moving the mouse after a book
expands.

Also, while I agree the shift is a little unusual, I think most people
whether new users or not will like it.